move smart gcal interface
