+++ modules/service/src/main/scala/lucuma/odb/logic/SmartGcalImplementation.scala
@@ -15,32 +15,13 @@ import lucuma.core.model.sequence.StepConfig
 import lucuma.core.model.sequence.StepConfig.Gcal
 import lucuma.core.model.sequence.gmos.DynamicConfig.GmosNorth
 import lucuma.core.model.sequence.gmos.DynamicConfig.GmosSouth
+import lucuma.odb.sequence.SmartGcalExpander
 import lucuma.odb.sequence.data.ProtoAtom
 import lucuma.odb.sequence.data.ProtoStep
 import lucuma.odb.service.SmartGcalService
 import lucuma.odb.smartgcal.data.Gmos
 
-trait SmartGcalExpander[F[_], K, D] {
-
-  /**
-   * Expands the given sequence, attempting to replace smart gcal steps with
-   * normal gcal steps.  For each atom in the Stream, if successful, the
-   * expanded atom without smart gcal steps is emitted in a `Right`.  When
-   * unsuccessful, the missing mapping is formatted and wrapped with a `Left`.
-   */
-  def expand: Pipe[F, ProtoAtom[ProtoStep[D]], Either[String, ProtoAtom[ProtoStep[D]]]]
-
-  def expandStep(
-    step: ProtoStep[D]
-  ): F[Either[String, NonEmptyList[ProtoStep[D]]]]
-
-  def expandAtom(
-    atom: ProtoAtom[ProtoStep[D]]
-  ): F[Either[String, ProtoAtom[ProtoStep[D]]]]
-
-}
-
-object SmartGcalExpander {
+object SmartGcalImplementation {
 
   type Cache[K, D] = Map[(K, SmartGcalType), Either[String, NonEmptyList[(D => D, Gcal)]]]
 
@@ -56,9 +37,9 @@ object SmartGcalExpander {
     toKey:     D => K,
     formatKey: K => String,
     select:    (K, SmartGcalType) => F[List[(D => D, Gcal)]]
-  ) extends SmartGcalExpander[F, K, D] {
+  ) extends SmartGcalExpander[F, D] {
 
-    val expand: Pipe[F, ProtoAtom[ProtoStep[D]], Either[String, ProtoAtom[ProtoStep[D]]]] =
+    override val expandSequence: Pipe[F, ProtoAtom[ProtoStep[D]], Either[String, ProtoAtom[ProtoStep[D]]]] =
       _.evalMapAccumulate(emptyCache[K, D])(expandAtomWithCache).map(_._2)
 
     private def expandAtomWithCache(
@@ -100,12 +81,12 @@ object SmartGcalExpander {
           (cache, NonEmptyList.one(ps).asRight).pure[F]
       }
 
-    def expandStep(
+    override def expandStep(
       step: ProtoStep[D]
     ): F[Either[String, NonEmptyList[ProtoStep[D]]]] =
       expandStepWithCache(emptyCache[K, D], step).map(_._2)
 
-    def expandAtom(
+    override def expandAtom(
       atom: ProtoAtom[ProtoStep[D]]
     ): F[Either[String, ProtoAtom[ProtoStep[D]]]] =
       expandAtomWithCache(emptyCache[K, D], atom).map(_._2)
@@ -114,14 +95,14 @@ object SmartGcalExpander {
 
   class ForInstrument[F[_]: Concurrent](service: SmartGcalService[F]) {
 
-    val gmosNorth: SmartGcalExpander[F, Gmos.SearchKey.North, GmosNorth] =
+    val gmosNorth: SmartGcalExpander[F, GmosNorth] =
       new Expander[F, Gmos.SearchKey.North, GmosNorth](
         Gmos.SearchKey.North.fromDynamicConfig,
         _.format,
         service.selectGmosNorth
       )
 
-    val gmosSouth: SmartGcalExpander[F, Gmos.SearchKey.South, GmosSouth] =
+    val gmosSouth: SmartGcalExpander[F, GmosSouth] =
       new Expander[F, Gmos.SearchKey.South, GmosSouth](
         Gmos.SearchKey.South.fromDynamicConfig,
         _.format,
